bpf: Temporary revert to Linux 4.7 version

+++ net/packet/af_packet.c
@@ -2434,7 +2434,7 @@ static void tpacket_set_protocol(const struct net_device *dev,
 
 static int __packet_snd_vnet_parse(struct virtio_net_hdr *vnet_hdr, size_t len)
 {
-	unsigned int gso_type = 0;
+	unsigned short gso_type = 0;
 
 	if ((vnet_hdr->flags & VIRTIO_NET_HDR_F_NEEDS_CSUM) &&
 	    (__virtio16_to_cpu(vio_le(), vnet_hdr->csum_start) +
